requirements.txt
dnspython
Flask==3.0.0
gunicorn